Reorganized Tabs exporting
