+++ src/components/Tabs/index.js
@@ -1,1 +1,5 @@
-export { default } from './Tabs';
+export { default as Tab } from './Tab';
+export { default as TabContent } from './TabContent';
+export { default as TabItem } from './TabItem';
+export { default as TabMenu } from './TabMenu';
+export { default as Tabs } from './Tabs';
