+++ src/index.js
@@ -13,10 +13,6 @@ export { default as LoadingSpinner } from './components/LoadingSpinner';
 export { NavBar, NavItem } from './components/Navigation';
 export { default as ProgressBar } from './components/ProgressBar';
 export { default as RadioButton } from './components/RadioButton';
-export { default as Tab } from './components/Tabs/Tab';
-export { default as TabContent } from './components/Tabs/TabContent';
-export { default as TabItem } from './components/Tabs/TabItem';
-export { default as TabMenu } from './components/Tabs/TabMenu';
-export { default as Tabs } from './components/Tabs';
+export { Tabs, Tab, TabContent, TabItem, TabMenu } from './components/Tabs';
 export { default as Text } from './components/Text';
 export { default as TextArea } from './components/TextArea';
